# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/namaz-times/my-masjid": {
      "get": {
        "operationId": "NamazTimesController_findMyMasjid",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Namaz times retrieved successfully"
          },
          "403": {
            "description": "Requires permission: namaz_times.read",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "message": "You are not allowed to access this masjid",
                    "errorCode": "MASJID_ACCESS_FORBIDDEN"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the current user's masjid namaz times",
        "tags": [
          "Namaz Times"
        ]
      },
      "put": {
        "operationId": "NamazTimesController_upsertMyMasjid",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertNamazTimeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Namaz times saved successfully"
          },
          "403": {
            "description": "Requires permission: namaz_times.update",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "message": "You are not allowed to access this masjid",
                    "errorCode": "MASJID_ACCESS_FORBIDDEN"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Insert or update the current user's masjid namaz times",
        "tags": [
          "Namaz Times"
        ]
      }
    },
    "/namaz-times/{masjidId}": {
      "get": {
        "operationId": "NamazTimesController_findByMasjidId",
        "parameters": [
          {
            "name": "masjidId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "4e0798d2-3fd3-4caa-9966-9f85c96f8b2f",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Namaz times retrieved successfully"
          },
          "403": {
            "description": "Requires permission: namaz_times.read",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "message": "You are not allowed to access this masjid",
                    "errorCode": "MASJID_ACCESS_FORBIDDEN"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "message": "You are not allowed to access this masjid",
                    "errorCode": "MASJID_ACCESS_FORBIDDEN"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get namaz times by masjid id",
        "tags": [
          "Namaz Times"
        ]
      },
      "put": {
        "operationId": "NamazTimesController_upsert",
        "parameters": [
          {
            "name": "masjidId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "4e0798d2-3fd3-4caa-9966-9f85c96f8b2f",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertNamazTimeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Namaz times saved successfully"
          },
          "403": {
            "description": "Requires permission: namaz_times.update",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "message": "You are not allowed to access this masjid",
                    "errorCode": "MASJID_ACCESS_FORBIDDEN"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "message": "You are not allowed to access this masjid",
                    "errorCode": "MASJID_ACCESS_FORBIDDEN"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Insert or update namaz times by masjid id",
        "tags": [
          "Namaz Times"
        ]
      }
    },
...
  },
  "info": {
    "title": "Masjid Core API",
    "description": "Backend for the masjid management system. Every masjid route needs a permission; see src/access/permissions.ts.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token from /auth/login/verify-otp"
      }
    },
    "schemas": {
...
      "UpsertNamazTimeDto": {
        "type": "object",
        "properties": {
          "fajr": {
            "type": "string",
            "example": "05:00 AM",
            "maxLength": 20
          },
          "zuhr": {
            "type": "string",
            "example": "01:30 PM",
            "maxLength": 20
          },
          "asr": {
            "type": "string",
            "example": "05:00 PM",
            "maxLength": 20
          },
          "maghrib": {
            "type": "string",
            "example": "06:45 PM",
            "maxLength": 20
          },
          "isha": {
            "type": "string",
            "example": "08:15 PM",
            "maxLength": 20
          },
          "jumma": {
            "type": "string",
            "example": "01:15 PM",
            "maxLength": 20
          },
          "note": {
            "type": "string",
            "example": "Jumma time may change in Ramadan",
            "maxLength": 500
          }
        }
      },
...
    }
  }
}