# masjid-core/src/modules/masjid-core/namaz-times/namaz-times.controller.ts
import {
  Body,
  Controller,
  Get,
  HttpStatus,
  Param,
  ParseUUIDPipe,
  Put,
  Req,
  UseGuards,
} from '@nestjs/common';
import {
  ApiBearerAuth,
  ApiBody,
  ApiOperation,
  ApiParam,
  ApiResponse,
  ApiTags,
} from '@nestjs/swagger';
import { JwtAuthGuard } from '../../platform-core/auth/guards/jwt-auth.guard';
import { AuthenticatedUser } from '../../platform-core/auth/types/jwt-payload.type';
import { UpsertNamazTimeDto } from './dto/upsert-namaz-time.dto';
import { NamazTimesService } from './namaz-times.service';
import { PERMISSIONS } from '../../../access/permissions';
import { RequirePermissions } from '../../../access/require-permissions';

type AuthenticatedRequest = {
  user: AuthenticatedUser;
};

const standardErrorSchema = {
  example: {
    success: false,
    message: 'You are not allowed to access this masjid',
    errorCode: 'MASJID_ACCESS_FORBIDDEN',
  },
};

@ApiTags('Namaz Times')
@ApiBearerAuth('bearer')
@UseGuards(JwtAuthGuard)
@Controller('namaz-times')
export class NamazTimesController {
  constructor(private readonly namazTimesService: NamazTimesService) {}

  // The my-masjid routes must stay above the :masjidId routes.
  @Get('my-masjid')
  @RequirePermissions(PERMISSIONS.NAMAZ_TIMES_READ)
  @ApiOperation({ summary: "Get the current user's masjid namaz times" })
  @ApiResponse({
    status: HttpStatus.OK,
    description: 'Namaz times retrieved successfully',
  })
  @ApiResponse({ status: HttpStatus.FORBIDDEN, schema: standardErrorSchema })
  findMyMasjid(@Req() request: AuthenticatedRequest) {
    return this.namazTimesService.findMyMasjid(request.user);
  }

  @Put('my-masjid')
  @RequirePermissions(PERMISSIONS.NAMAZ_TIMES_UPDATE)
  @ApiOperation({
    summary: "Insert or update the current user's masjid namaz times",
  })
  @ApiBody({ type: UpsertNamazTimeDto })
  @ApiResponse({
    status: HttpStatus.OK,
    description: 'Namaz times saved successfully',
  })
  @ApiResponse({ status: HttpStatus.FORBIDDEN, schema: standardErrorSchema })
  upsertMyMasjid(
    @Body() dto: UpsertNamazTimeDto,
    @Req() request: AuthenticatedRequest,
  ) {
    return this.namazTimesService.upsertMyMasjid(dto, request.user);
  }

  @Get(':masjidId')
  @RequirePermissions(PERMISSIONS.NAMAZ_TIMES_READ)
  @ApiOperation({ summary: 'Get namaz times by masjid id' })
  @ApiParam({
    name: 'masjidId',
    example: '4e0798d2-3fd3-4caa-9966-9f85c96f8b2f',
  })
  @ApiResponse({
    status: HttpStatus.OK,
    description: 'Namaz times retrieved successfully',
  })
  @ApiResponse({ status: HttpStatus.FORBIDDEN, schema: standardErrorSchema })
  @ApiResponse({ status: HttpStatus.NOT_FOUND, schema: standardErrorSchema })
  findByMasjidId(
    @Param('masjidId', new ParseUUIDPipe({ version: '4' })) masjidId: string,
    @Req() request: AuthenticatedRequest,
  ) {
    return this.namazTimesService.findByMasjidId(masjidId, request.user);
  }

  @Put(':masjidId')
  @RequirePermissions(PERMISSIONS.NAMAZ_TIMES_UPDATE)
  @ApiOperation({ summary: 'Insert or update namaz times by masjid id' })
  @ApiParam({
    name: 'masjidId',
    example: '4e0798d2-3fd3-4caa-9966-9f85c96f8b2f',
  })
  @ApiBody({ type: UpsertNamazTimeDto })
  @ApiResponse({
    status: HttpStatus.OK,
    description: 'Namaz times saved successfully',
  })
  @ApiResponse({ status: HttpStatus.FORBIDDEN, schema: standardErrorSchema })
  @ApiResponse({ status: HttpStatus.NOT_FOUND, schema: standardErrorSchema })
  upsert(
    @Param('masjidId', new ParseUUIDPipe({ version: '4' })) masjidId: string,
    @Body() dto: UpsertNamazTimeDto,
    @Req() request: AuthenticatedRequest,
  ) {
    return this.namazTimesService.upsert(masjidId, dto, request.user);
  }
}